% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarks.R
\name{clusterKmers}
\alias{clusterKmers}
\title{Cluster k-mers}
\usage{
clusterKmers(kmers, k = 4, nClusters = NULL)
}
\arguments{
\item{kmers}{character vector of k-mers to partition into clusters}

\item{k}{length of sub-k-mers (default k=4 to use tetramers) with
which to calculate Jaccard distances for clustering}

\item{nClusters}{number of clusters to partition kmers into; if set
to NULL (default value), selects number of clusters to maximize
the average silhouette score
(\url{https://en.wikipedia.org/wiki/Silhouette_(clustering)}).}
}
\description{
Takes a set of k-mer sequences and returns a list of partitioning
the input k-mers into clusters of more similar k-mers. Hierarchical
clustering (average linkage) is performed based on Jaccard
coefficient distance metric applied treating each k-mer as the set
of all tetramers which can be found as substrings within it.
Tetramers which are reverse-complements of each other are treated
as equivalent.
}
