% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarks.R
\name{dnaRevComp}
\alias{dnaRevComp}
\title{Reverse-complement DNA sequence}
\usage{
dnaRevComp(s)
}
\arguments{
\item{s}{character vector DNA sequence(s)}
}
\value{
character vector containing reverse-complement(s) of s
}
\description{
Takes character vector input representing DNA sequence and returns
reverse-complement as character vector.
}
\examples{
dnaRevComp(c('TGACCTT', 'CATACTGAGA'))

}
