% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarks.R
\name{locateClusters}
\alias{locateClusters}
\title{Locate occurrences of specified clusters of k-mers}
\usage{
locateClusters(clusters, seqs, directional = TRUE)
}
\arguments{
\item{clusters}{list of character vectors of k-mers to search for}

\item{seqs}{character vector of sequences in which to locate kmer}

\item{directional}{logical value: if FALSE, counts occurrences of
either k-mers within each cluster or their reverse-complements.}
}
\value{
data.frame with three columns: `seqid' containing the name
    of the sequence from seqs in which the match was found;
    `cluster' indicating the cluster from wich a k-mer was located;
    and `location' giving the 1-based position at which the match
    was found.
}
\description{
Find locations of matches of list of character vectors of k-mers in
each element of a character vector.
}
