% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulatedScores}
\alias{simulatedScores}
\title{Scores associated with simulated sequences from SArKS paper.}
\format{Named numeric vector.}
\source{
\url{https://github.com/denniscwylie/sarks/tree/master/examples/simulated_scores.tsv}
}
\usage{
simulatedScores
}
\description{
Scores associated with simulated DNA sequences used to illustrate suffix array kernel smoothing method in \url{https://academic.oup.com/bioinformatics/article-abstract/35/20/3944/5418797} (first 20 sequences assigned score of 0.0, last 10 assigned score of 1.0).
}
\keyword{datasets}
