% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulatedSeqs}
\alias{simulatedSeqs}
\title{Simulated sequences from SArKS paper.}
\format{Named character vector.}
\source{
\url{https://github.com/denniscwylie/sarks/tree/master/examples/simulated_seqs.fa}
}
\usage{
simulatedSeqs
}
\description{
Simulated DNA sequences used to illustrate suffix array kernel
smoothing method in
\url{https://academic.oup.com/bioinformatics/article-abstract/35/20/3944/5418797}.
}
\keyword{datasets}
