% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sarks.R
\name{slurpFasta}
\alias{slurpFasta}
\title{Slurp fasta file into character vector}
\usage{
slurpFasta(f)
}
\arguments{
\item{f}{the path for the fasta file to read in}
}
\value{
a named character vector: names are the sequence names from
    the fasta file, values are the sequences themselves.
}
\description{
Convenience function for quickly reading a fasta file into an R character vector.
}
